Updating Qthreads implementation of Chapel interface to reflect recent changes
This update catches the qthreads tree up with recent changes in the
Chapel tasking interface:

* removes the chpl_task_getMax*() routines which have been deprecated

* updates the chpl_task_init() routine which now takes the requested
  number of threads (shepherds) an upper limit on the number of
  threads/shepherds if one exists (e.g., as imposed by the
  communication layer), and whether or not a special task is going to
  be requested for communication progress.

  - updated the logic of this routine to create the requested number
    of shepherds, and to use a number equal to the number of cores
    on the locale (as determined by Chapel's utility function) if it
    is unspecified.

  - also changed the default callstack size to use something larger
    than qthreads does by default in order to make it more similar
    to other tasking layers/less likely to seg fault for nontrivial
    programs.


Making these same changes to the snapshot of qthreads in Chapel's SVN
repository has not hurt stability and has made testing work better
(primarily due to the callstack size change).  Unfortunately, I'm
getting instabilities when using the trunk version of qthreads with
the trunk version of Chapel, but have not had a chance to diagnose yet
(and may not until we consider a new snapshot of qthreads in hopes
that it stabilizes itself).





git-svn-id: https://qthreads.googlecode.com/svn/trunk@2333 9dbd018f-d1da-d342-4c11-58c73dfe904b