Optional profiling of the distribution of retries on compare-and-swap operations in the nottingham scheduler. Enable with -DCAS_STEAL_PROFILE (./configure option is forthcoming)
A stripe of integer counters is used to track
the number of CAS retries. Each qthread worker
is assigned its own stripe, and the stripes are
tallied at the end of computation.

This is a time/space tradeoff to avoid contention
for shared counters.

(qt_threadqueue.h) forward declaration for print statistics function
(qthread_innards.h) add the array of counter stripes
(nottingham_threadqueues.c) monitor push() and pop() operations
(qthread.c) initialization and teardown of array of counter stripes


git-svn-id: https://qthreads.googlecode.com/svn/trunk@2751 9dbd018f-d1da-d342-4c11-58c73dfe904b