Initial commit of the loxley scheduler.
The loxley scheduler is nearly identical to sherwood, in that both schedulers are protected by a giant lock. However, the loxley stack is implemented as a contiguous region of memory, whereas the sherwood stack is implemented as a linked list. We intend to use loxley as a testing framework for various extensions that will reduce contention on the shared stack.


git-svn-id: https://qthreads.googlecode.com/svn/trunk@2786 9dbd018f-d1da-d342-4c11-58c73dfe904b