Contention management on loxley scheduler.
The loxley scheduler now has an array of local stacks
that are used as a backup in case of contention on the
primary stacks. Contention is determined indirectly from
a failure to acquire a tryLock() protecting the shared stacks.

Spinning mutexes with a trylock() functionality have been 
added to qt_atomics.h. If QTHREAD_ATOMIC_INCR is defined,
then a fast implementation is used. Otherwise the pthreads
implementation is used as a fallback.



git-svn-id: https://qthreads.googlecode.com/svn/trunk@2787 9dbd018f-d1da-d342-4c11-58c73dfe904b