Hacks the build system to pass single-worker scheduler tests.
This is required to get the nightly tests to pass for single-worker
schedulers. The tests are failing because the affinity layers attempt to
use more than one worker per shepherd and an error condition is hit when
the Qthreads detects that the threadqueue and system topology are not
compatible.

The right way to do this is to update all of the affinity layer
implementations recognize when there is a constraint that only one
worker is supported per shepherd.