Limit the pool max alloc_size.
The existig code worked fine when all pools contained smallish objects.
But with Chapel the default stack size is 8 mb and at least the stack
pool could end up quite large, especially when the stack size wasn't a
multiple of the page size and guard pages (which would have forced the
size to a whole number of pages) were not used.

The qt_gcd.h change is related, in that testing demonstrated incorrect
results for qt_lcm() with larger values on 32-bit systems.