Add a writeFF queue (FFWQ) in preparation for adding writeFF support
This just adds a new queue to be used for writeFF support. I based it off the
readFF queue (FFQ) with the minor exception of performing a write operation
instead of a read in qthread_gotlock_fill_inner()

The only use of FFQ that I wasn't sure about was the one in
qthread_check_feb_preconds(), but this just appears to be a modified version of
readFF that doesn't block, so I don't think I need to do anything with this.

Note that I added support for FFQ in feb.c, I have no current plans on adding
support for the syncvar.c implementation.