Add initial qthread_writeFF
Add a WRONG version of qthread_writeFF. It's actually just a renamed copy of
readFF so that it'll be easy to see the differences between readFF and writeFF
in the next patch. i.e. writeFF is a cross of write* and readFF so I wanted to
start with a readFF base so that it's easier to see in the diff that I'm just
adding a readFF with the read parts swapped to write, but that the
locking/unlocking logic is the same.