Add initial qthread_writeE
Add a WRONG version of qthread_writeE. It's actually just a renamed copy of
qthread_empty so that it'll be easy to see the differences between writeE and
empty in the next patch.

i.e. writeE() is really just empty() plus a write operation, so I'm starting
with empty() so that it's easier to see in the diff that the locking/unlocking
logic is the same and I'm really just adding a write operation.