Add qthread_readXX
Add support for qthread_readXX. This is really just a simple wrapper around

```c
if (dest && (dest != src)) {
    *(aligned_t *)dest = *(aligned_t *)src;
    MACHINE_FENCE;
}
```

but it includes debug and FEB profiling calls. Note that the code that gets the
current shepherd could have been wrapped in ifdefs for QTHREAD_FEB_PROFILING
and QTHREAD_DEBUG to avoid some unnecessary work for production builds, but it
didn't really seem worthwhile to complicate the code to save a few operations
that are relatively cheap in the context of other FEB operations.