Support Portals4
This revision updates multinode Qthreads to support the modern Portals4
Runtime detailed here:
https://github.com/portals4/portals4/commit/ff751cc53226dc1365f02fc7831cfb8b5f3cbb48

Since Portals4 no longer supports multiple runtimes (the only runtime is
libportals_runtime) I removed the --with-multinode-runtime option
entirely. To enable multinode all you need is the following configure
invocation:

./configure --enable-multinode --with-portals4=<dir>

I also changed the "OMPI_CHECK_PACKAGE" macro to support cases where the
include files of a package are located in the root of the library
directory rather than in an include directory relative to the base. This
is necessary to support the Portals4 runtime.

Finally I fixed a race in the multinode test case in_edges.c by adding a
barrier. This fixes a deadlock in the test cases and allows all test
cases to pass on my system.