External allocators
This is an initial draft of a framework to add an external allocator API to
qthreads. You can select an external allocator framework with the
--with-alloc flag. Right now the only allocator available is "base"
which just maps to the traditional C malloc. It passes the usual
qthreads tests but may have other bugs.

Right now there are two allocators, base and chapel.