Initial fix of sleep system call and new test case
The marshaled arguments of the sleep system call in the system call
forwarding layer was wrong. Add timespec structures to call the functions
correctly and add a test case to exercise the call to sleep (this will
need some work to be a genuine test, right now it's just a way of eyeballing).