Thread sanitizer interop to follow execution order constraints across context swaps.
Note: make stored tsan fibers atomic to clear a tsan error. This makes some sense because the fiber entry has to be accessed in order to prepare or finalize the context swap and those operations may happen in different OS threads when qthreads get stolen or just suspended for the IO subsystem.